Stop sessions leaking between Behat scenarios
There was an issue whereby Mink sessions were leaking between Scenarios.

This was problematic when using Wordhat's 'Given I am logged in as a [role]' step, which will attempt to log out if the body element has the class '.logged-in'. Its log out method involves clicking the 'Logout' link on the admin toolbar. However subscriber users do not see the admin toolbar, and so it was impossible for Wordhat to log out of a subscriber session.

This resolves the issue by simply resetting the session and navigating to the homepage at the end of each session. This will cause the page to reload, and the new body element will no longer have a '.logged-in' class, thus Wordhat will no longer attempt to log out.